added descriptions, bug fixes and more
Added dictionaries for the descriptions of the characters and the locations, accessible through the main menu. Fixed a bug where if you type quit during the selection of weapons in the combat section it would say invalid action. more detailed and acurrate comments and docstrings are provided for some parts of the code. Added file extension in the title